+++ packages/svelte/tests/runtime-runes/samples/form-default-value-from-spread/_config.js
@@ -1,0 +1,10 @@
+import { test } from '../../test';
+
+export default test({
+	mode: ['server'],
+	html: `
+		<input value="a">
+		<input type="checkbox" checked>
+		<input value="b">
+`
+});