+++ packages/svelte/tests/snapshot/samples/each-index-non-null/_expected/server/index.svelte.js
@@ -1,13 +1,13 @@
 import * as $ from 'svelte/internal/server';
 
 export default function Each_index_non_null($$payload) {
+	$$payload.push(`<!--[-->`);
+
 	const each_array = $.ensure_array_like(Array(10));
 
-	$$payload.out.push(`<!--[-->`);
-
 	for (let i = 0, $$length = each_array.length; i < $$length; i++) {
-		$$payload.out.push(`<p>index: ${$.escape(i)}</p>`);
+		$$payload.push(`<p>index: ${$.escape(i)}</p>`);
 	}
 
-	$$payload.out.push(`<!--]-->`);
+	$$payload.push(`<!--]-->`);
 }