+++ packages/svelte/tests/snapshot/samples/each-string-template/_expected/server/index.svelte.js
@@ -1,15 +1,15 @@
 import * as $ from 'svelte/internal/server';
 
 export default function Each_string_template($$payload) {
-	const each_array = $.ensure_array_like(['foo', 'bar', 'baz']);
+	$$payload.push(`<!--[-->`);
 
-	$$payload.out.push(`<!--[-->`);
+	const each_array = $.ensure_array_like(['foo', 'bar', 'baz']);
 
 	for (let $$index = 0, $$length = each_array.length; $$index < $$length; $$index++) {
 		let thing = each_array[$$index];
 
-		$$payload.out.push(`<!---->${$.escape(thing)}, `);
+		$$payload.push(`<!---->${$.escape(thing)}, `);
 	}
 
-	$$payload.out.push(`<!--]-->`);
+	$$payload.push(`<!--]-->`);
 }